Issue-54 move ImmutableCollectors to util-java8 project
