+++ server/pom.xml
@@ -67,6 +67,7 @@
         <module>container/mailbox-adapter</module>
         <module>container/spring</module>
         <module>container/util</module>
+        <module>container/util-java8</module>
 
         <module>mailet/mailetcontainer-api</module>
         <module>mailet/mailetcontainer-camel</module>
@@ -671,6 +672,11 @@
             <dependency>
                 <groupId>org.apache.james</groupId>
                 <artifactId>james-server-mailets</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.james</groupId>
+                <artifactId>james-server-util-java8</artifactId>
                 <version>${project.version}</version>
             </dependency>
